backend/requirements.txt
# Base CLI
click==8.1.3

# Visualization & EDA
matplotlib==3.7.5
seaborn==0.13.2
statsmodels==0.14.1

# Data processing
numpy==1.24.4
pandas==2.0.3
scikit_learn==1.3.2

# MLflow tracking
mlflow==2.19.0
mlflow-skinny==2.19.0

# ZenML stack
zenml
zenml[server]

# Web backend
Flask==2.3.2
Flask-Cors==4.0.0

# Utilities (if not already installed)
python-dotenv==1.0.1


# for mongodb
pymongo==4.5.0
dnspython==2.3.0

# test 
pytest
werkzeug==3.0.1


mongomock